+++ govee_api_laggat/govee_api_laggat.py
@@ -1,13 +1,14 @@
 """ client to connect to the govee API """
 
-import sys
-import logging
-import time
-from datetime import datetime
-import asyncio
+from govee_api_laggat.learning_storage import GoveeAbstractLearningStorage, GoveeLearnedInfo
 import aiohttp
+import asyncio
 from dataclasses import dataclass
-from typing import List, Tuple, Union, Any
+from datetime import datetime
+import logging
+import sys
+import time
+from typing import List, Tuple, Union, Optional, Any
 
 _LOGGER = logging.getLogger(__name__)
 _API_URL = "https://developer-api.govee.com"
@@ -15,8 +16,9 @@ _API_URL = "https://developer-api.govee.com"
 _RATELIMIT_TOTAL = 'Rate-Limit-Total' # The maximum number of requests you're permitted to make per minute.
 _RATELIMIT_REMAINING = 'Rate-Limit-Remaining' # The number of requests remaining in the current rate limit window.
 _RATELIMIT_RESET = 'Rate-Limit-Reset' # The time at which the current rate limit window resets in UTC epoch seconds.
-BRIGHTNESS_254_MODELS = [] # maybe these? support asked: ["H6089","H7022","H6086","H6135","H6137","H7005","H6002","H6003"]
 
+# return state from hisory for n seconds after controlling the device
+DELAY_GET_FOLLOWING_SET_SECONDS = 2
 @dataclass
 class GoveeDevice(object):
     """ Govee Device DTO """
@@ -30,19 +32,18 @@ class GoveeDevice(object):
     support_brightness: bool
     support_color: bool
     support_color_tem: bool
-    
-@dataclass
-class GoveeDeviceState(object):
-    """ State of a Govee Device DTO """
-    device: str
-    model: str
     online: bool
     power_state: bool
     brightness: int
     color: Tuple[ int, int, int ]
     timestamp: int
     source: str
-
+    error: str
+    lock_set_until: int
+    lock_get_until: int
+    learned_set_brightness_max: int
+    learned_get_brightness_max: int
+    
 class Govee(object):
     """ client to connect to the govee API """
 
@@ -57,20 +58,28 @@ class Govee(object):
             await self._session.close()
         self._session = None
     
-    def __init__(self, api_key: str):
+    def __init__(self, api_key: str, *, 
+            learning_storage: Optional[GoveeAbstractLearningStorage] = None
+        ):
         """ init with an API_KEY """
         self._api_key = api_key
         self._devices = {}
-        self._states = {}
         self._rate_limit_on = 5 # safe available call count for multiple processes
         self._limit = 100
         self._limit_remaining = 100
         self._limit_reset = 0
-        self._no_state_before = 0
+        self._learning_storage = learning_storage
+        if not self._learning_storage:
+            # use an internal learning storage as long as we run.
+            # we will need to re-learn every time again.
+            self._learning_storage = GoveeAbstractLearningStorage()
+
         
     @classmethod
-    async def create(cls, api_key: str):
-        self = Govee(api_key)
+    async def create(cls, api_key: str, *, 
+            learning_storage: Optional[GoveeAbstractLearningStorage] = None
+        ):
+        self = Govee(api_key, learning_storage=learning_storage)
         await self.__aenter__()
         return self
     
@@ -82,12 +91,6 @@ class Govee(object):
 
     def _utcnow(self):
         return datetime.timestamp(datetime.now())
-
-    def _no_state_request_next_seconds(self, sec: int):
-        self._no_state_before = self._utcnow() + sec
-    
-    def _state_request_allowed(self) -> bool:
-        return self._no_state_before < self._utcnow()
 
     def _track_rate_limit(self, response):
         """ rate limiting information """
@@ -101,7 +104,7 @@ class Govee(object):
                 self._limit_reset = float(response.headers[_RATELIMIT_RESET])
                 _LOGGER.debug(f'Rate limit total: {self._limit}, remaining: {self._limit_remaining} in {self.rate_limit_reset_seconds} seconds')
                 limit_unknown = False
-            except ex:
+            except Exception as ex:
                 _LOGGER.warning(f'Error trying to set rate limits: {ex}')
         if limit_unknown:
             self._limit_remaining = 0
@@ -146,33 +149,23 @@ class Govee(object):
         self._rate_limit_on = val
     
     @property
-    def devices(self):
+    def devices(self) -> List[GoveeDevice]:
         """ returns the cached devices list """
         lst = []
         for dev in self._devices:
             lst.append(self._devices[dev])
         return lst
     
-    @property
-    def states(self):
-        """ returns the cached states list """
-        lst = []
-        for dev in self._states:
-            lst.append(self._states[dev])
-        return lst
-    
-    def state(self, device) -> GoveeDeviceState:
-        """ returns the cached state for a device """
-        device_str, device = self._get_device(device)
-        if device_str:
-            return self._states[device_str]
-        return None
+    def device(self, device) -> GoveeDevice:
+        """ returns the cached device """
+        _, device = self._get_device(device)
+        return device
 
-    async def ping_async(self) -> Tuple[ float, str ]:
+    async def ping(self) -> Tuple[ float, str ]:
         """ Ping the api endpoint. No API_KEY is needed
             Returns: timeout_ms, error
         """
-        _LOGGER.debug("ping_async")
+        _LOGGER.debug("ping")
         start = time.time()
         ping_ok_delay = None
         err = None
@@ -180,7 +173,7 @@ class Govee(object):
         url = (_API_URL + "/ping")
         await self.rate_limit_delay()
         async with self._session.get(url=url) as response:
-            self._track_rate_limit(response)
+            # no rate limit header fields exist on ping
             result = await response.text()
             delay = int((time.time() - start) * 1000)
             if response.status == 200:
@@ -193,12 +186,10 @@ class Govee(object):
                 err = f'API-Error {response.status}: {result}'
         return ping_ok_delay, err
         
-            
     async def get_devices(self) -> Tuple[ List[GoveeDevice], str ]:
         """ get and cache devices, returns: list, error """
         _LOGGER.debug("get_devices")
         devices = {}
-        states = {}
         err = None
         
         url = (
@@ -209,13 +200,27 @@ class Govee(object):
         async with self._session.get(url=url, headers = self._getAuthHeaders()) as response:
             self._track_rate_limit(response)
             if response.status == 200:
-                self._no_state_request_next_seconds(2)
                 result = await response.json()
                 timestamp = self._utcnow()
                 
+                learning_infos = await self._learning_storage._read_cached()
+
                 for item in result["data"]["devices"]:
-                    devices[item["device"]] = GoveeDevice(
-                        device = item["device"],
+                    device_str = item["device"]
+
+                    # assuming max values for control and feedback of brightness
+                    learned_set_brightness_max = None
+                    learned_get_brightness_max = None
+                    if device_str in learning_infos:
+                        learning_info = learning_infos[device_str]
+                        learned_set_brightness_max = learning_info.set_brightness_max
+                        learned_get_brightness_max = learning_info.get_brightness_max
+                    if not item["retrievable"]:
+                        learned_get_brightness_max = -1
+
+                    # create device DTO
+                    devices[device_str] = GoveeDevice(
+                        device = device_str,
                         model = item["model"],
                         device_name = item["deviceName"],
                         controllable = item["controllable"],
@@ -224,24 +229,25 @@ class Govee(object):
                         support_turn = "turn" in item["supportCmds"],
                         support_brightness = "brightness" in item["supportCmds"],
                         support_color = "color" in item["supportCmds"],
-                        support_color_tem = "colorTem" in item["supportCmds"]
-                    )
-                    states[item["device"]] = GoveeDeviceState(
-                        device = item["device"],
-                        model = item["model"],
+                        support_color_tem = "colorTem" in item["supportCmds"],
+                        # defaults for state
                         online = True,
                         power_state = False,
                         brightness = 0,
                         color = (0, 0, 0), 
                         timestamp = timestamp,
-                        source = 'history'
+                        source = 'history',
+                        error = None,
+                        lock_set_until = 0,
+                        lock_get_until = 0,
+                        learned_set_brightness_max = learned_set_brightness_max,
+                        learned_get_brightness_max = learned_get_brightness_max,
                     )
             else:
                 result = await response.text()
                 err = f'API-Error {response.status}: {result}'
         # cache last get_devices result
         self._devices = devices
-        self._states = states
         return self.devices, err
 
 
@@ -281,9 +287,9 @@ class Govee(object):
             if not err:
                 success = self._is_success_result_message(result)
                 if success:
-                    self._states[device_str].timestamp = self._utcnow
-                    self._states[device_str].source = 'history'
-                    self._states[device_str].power_state = onOff == "on"
+                    self._devices[device_str].timestamp = self._utcnow
+                    self._devices[device_str].source = 'history'
+                    self._devices[device_str].power_state = onOff == "on"
         return success, err
 
     async def set_brightness(self, device: Union[str, GoveeDevice], brightness: int) -> Tuple[ bool, str ]:
@@ -299,21 +305,56 @@ class Govee(object):
             else:
                 # set brightness as 0..254
                 brightness_set = brightness
-                if device.model not in BRIGHTNESS_254_MODELS:
+                brightness_set_100 = brightness * 100 // 254
+                if device.learned_set_brightness_max == 100:
                     # set brightness as 0..100
-                    brightness_set = brightness * 100 // 254
+                    brightness_set = brightness_set_100
                 command = "brightness"
                 result, err = await self._control(device, command, brightness_set)
+                if err:
+                    # try again with 0-100 range
+                    if device.learned_set_brightness_max == None and "API-Error 400" in err:
+                        # set brightness as 0..100 as 0..254 didn't work
+                        brightness_set = brightness_set_100
+                        result, err = await self._control(device, command, brightness_set)
+                        if not err:
+                            device.learned_set_brightness_max = 100
+                            await self._learn(device)
+                else:
+                    device.learned_set_brightness_max = 254
+                    await self._learn(device)
+
                 if not err:
                     success = self._is_success_result_message(result)
                     if success:
-                        self._states[device_str].timestamp = self._utcnow
-                        self._states[device_str].source = 'history'
-                        self._states[device_str].brightness = brightness
+                        self._devices[device_str].timestamp = self._utcnow
+                        self._devices[device_str].source = 'history'
+                        self._devices[device_str].brightness = brightness
+                        self._devices[device_str].power_state = brightness == 0
         return success, err
 
+    async def _learn(self, device):
+        """Persist learned information from device DTO."""
+        learning_infos: Dict[str, GoveeLearnedInfo] = await self._learning_storage._read_cached()
+        # init Dict and entry for device
+        if learning_infos == None:
+            learning_infos = {}
+        if device.device not in learning_infos:
+            learning_infos[device.device] = GoveeLearnedInfo()
+        # output what was lerned, and learn
+        if learning_infos[device.device].set_brightness_max != device.learned_set_brightness_max:
+            _LOGGER.debug("learned device %s uses range 0-%s for setting brightness.", device.device, device.learned_set_brightness_max)
+            learning_infos[device.device].set_brightness_max = device.learned_set_brightness_max
+        if learning_infos[device.device].get_brightness_max != device.learned_get_brightness_max:
+            _LOGGER.debug("learned device %s uses range 0-%s for getting brightness state.", device.device, device.learned_get_brightness_max)
+            if device.learned_get_brightness_max == 100:
+                _LOGGER.info("brightness range for %s is assumed. If the brightness slider doesn't match the actual brightness pull the brightness up to max once.", device.device)
+            learning_infos[device.device].get_brightness_max = device.learned_get_brightness_max
+        
+        await self._learning_storage._write_cached(learning_infos)
+
     async def set_color_temp(self, device: Union[str, GoveeDevice], color_temp: int) -> Tuple[ bool, str ]:
-        """ set color temperature to 2000 .. 9000 """
+        """ set color temperature to 2000 .. 9000."""
         success = False
         err = None
         device_str, device = self._get_device(device)
@@ -328,9 +369,9 @@ class Govee(object):
                 if not err:
                     success = self._is_success_result_message(result)
                     if success:
-                        self._states[device_str].timestamp = self._utcnow
-                        self._states[device_str].source = 'history'
-                        self._states[device_str].color_temp = color_temp
+                        self._devices[device_str].timestamp = self._utcnow
+                        self._devices[device_str].source = 'history'
+                        self._devices[device_str].color_temp = color_temp
         return success, err
 
     async def set_color(self, device: Union[str, GoveeDevice], color: Tuple[ int, int, int ]) -> Tuple[ bool, str ]:
@@ -360,10 +401,16 @@ class Govee(object):
                     if not err:
                         success = self._is_success_result_message(result)
                         if success:
-                            self._states[device_str].timestamp = self._utcnow
-                            self._states[device_str].source = 'history'
-                            self._states[device_str].color = color
+                            self._devices[device_str].timestamp = self._utcnow
+                            self._devices[device_str].source = 'history'
+                            self._devices[device_str].color = color
         return success, err
+
+    def _get_lock_seconds(self, utcSeconds: int) -> int:
+        seconds_lock = utcSeconds - self._utcnow()
+        if(seconds_lock < 0):
+            seconds_lock = 0
+        return seconds_lock
 
     async def _control(self, device: Union[str, GoveeDevice], command: str, params: Any) -> Tuple[ Any, str ]:
         device_str, device = self._get_device(device)
@@ -377,8 +424,11 @@ class Govee(object):
         if not device:
             err = f'Invalid device {device_str}, {device}'
         else:
+            seconds_locked = self._get_lock_seconds(device.lock_set_until)
             if not device.controllable:
                 err = f'Device {device.device} is not controllable'
+            elif seconds_locked:
+                err = f'Device {device.device} is locked for control next {sec} seconds'
             elif not command in device.support_cmds:
                 err = f'Command {command} not possible on device {device.device}'
             else:
@@ -399,29 +449,40 @@ class Govee(object):
                 ) as response:
                     self._track_rate_limit(response)
                     if response.status == 200:
-                        self._no_state_request_next_seconds(2)
+                        device.lock_get_until = self._utcnow() + DELAY_GET_FOLLOWING_SET_SECONDS
                         result = await response.json()
                     else:
                         text = await response.text()
                         err = f'API-Error {response.status} on command {cmd}: {text} for device {device}'
         return result, err
 
-    async def get_state(self, device: Union[str, GoveeDevice]) -> Tuple[ GoveeDeviceState, str ]:
+    async def get_states(self) -> List[GoveeDevice]:
+        _LOGGER.debug('get_states')
+        for device_str in self._devices:
+            state, err = await self._get_device_state(device_str)
+            if err:
+                self._devices[device_str].error = err
+            else:
+                self._devices[device_str] = state
+                self._devices[device_str].error = None
+        return self.devices
+
+    async def _get_device_state(self, device: Union[str, GoveeDevice]) -> Tuple[ GoveeDevice, str ]:
         device_str, device = self._get_device(device)
-        _LOGGER.debug(f'get_state {device_str}')
         result = None
         err = None
+        seconds_locked = self._get_lock_seconds(device.lock_get_until)
         if not device:
             err = f'Invalid device {device_str}'
         elif not device.retrievable:
             # device {device_str} isn't able to return state, return 'history' state
-            self._states[device_str].source = 'history'
-            result = self._states[device_str]
-        elif not self._state_request_allowed():
+            self._devices[device_str].source = 'history'
+            result = self._devices[device_str]
+        elif seconds_locked:
             # we just changed something, return state from history
-            self._states[device_str].source = 'history'
-            result = self._states[device_str]
-            _LOGGER.debug(f'state object returned from cache: {result}')
+            self._devices[device_str].source = 'history'
+            result = self._devices[device_str]
+            _LOGGER.debug(f'state object returned from cache: {result}, next state for {device.device} from api allowed in {seconds_locked} seconds')
         else:
             url = (
                 _API_URL
@@ -463,20 +524,31 @@ class Govee(object):
                         else:
                             _LOGGER.warning(f'unknown state property {prop}')
 
-                    result = GoveeDeviceState(
-                        device = json_obj["data"]["device"],
-                        model = json_obj["data"]["model"],
-                        online = prop_online,
-                        power_state = prop_power_state,
-                        brightness = prop_brightness,
-                        color = prop_color, 
-                        timestamp = timestamp,
-                        source = 'api'
-                    )
-                    self._states[result.device] = result
+                    #autobrightness learning
+                    if device.learned_get_brightness_max == None \
+                        or ( \
+                            device.learned_get_brightness_max == 100 \
+                            and prop_brightness > 100 \
+                        ):
+                        device.learned_get_brightness_max = 100  # assumption, as we didn't get anything higher
+                        if prop_brightness > 100:
+                            device.learned_get_brightness_max = 254
+                        await self._learn(device)
+                    if device.learned_get_brightness_max == 100:
+                        # scale range 0-100 up to 0-254
+                        prop_brightness = prop_brightness * 254 // 100
+
+                    result = self._devices[device_str]
+                    result.online = prop_online
+                    result.power_state = prop_power_state
+                    result.brightness = prop_brightness
+                    result.color = prop_color
+                    result.timestamp = timestamp
+                    result.source = 'api'
+                    result.error = None
+
                     _LOGGER.debug(f'state returned from API: {json_obj}, resulting state object: {result}')
                 else:
                     errText = await response.text()
                     err = f'API-Error {response.status}: {errText}'
         return result, err
-